requirements.txt
numpy
python-dotenv
gradio==5.13.1
tqdm
soundfile
azure-identity